requirements.txt
asgiref==3.7.2
Django==4.2.6
sqlparse==0.4.4
typing_extensions==4.8.0
djangorestframework==3.14.0
firebase-admin==6.2.0
django-cors-headers==4.3.0
